hardwareBackend/requirements.txt
pyserial==3.5
requests==2.32.3
python-dotenv==1.0.1